refactor: remove utility function definitions from `_app.ts`
